+++ OUDS/Core/Themes/Orange/Sources/Providers/ComponentTokens/OrangeThemeBreadcrumbComponentTokensProvider.swift
@@ -64,6 +64,20 @@ import OUDSTokensSemantic
 ///     OrangeTheme(breadcrumb: CustomBreadcrumbComponentTokensProvider())
 /// ```
 ///
+/// It is also possible to use your own semantic tokens providers for this component tokens providers:
+///
+/// ```swift
+///     // Uses by default here:
+///     // - OrangeThemeSizeSemanticTokensProvider for sizes
+///     // - OrangeThemeSpaceSemanticTokensProvider for spaces
+///     let breadcrumbComponentTokensProvider = OrangeThemeBreadcrumbComponentTokensProvider()
+///
+///     // Or use your own size and space semantic tokens providers (or only some)
+///     let breadcrumbComponentTokensProvider = OrangeThemeBreadcrumbComponentTokensProvider(
+///                                                 sizes: CustomSizeSemanticTokensProvider(),
+///                                                 spaces: CustomSpaceSemanticTokensProvider())
+/// ```
+///
 /// - Since: 0.10.0
 open class OrangeThemeBreadcrumbComponentTokensProvider: AllBreadcrumbComponentTokensProvider {
 
@@ -75,13 +89,13 @@ open class OrangeThemeBreadcrumbComponentTokensProvider: AllBreadcrumbComponentT
 
     /// Defines a provider of component tokens dedicated to `OUDSLink`
     /// - Parameters:
-    ///    - sizes: Provider for size semantic tokens
-    ///    - spaces: Provider for space semantic tokens
-    public init(sizes: AllSizeSemanticTokensProvider,
-                spaces: AllSpaceSemanticTokensProvider) {
+    ///    - sizes: Provider for size semantic tokens. If nil, a default one will be used (``OrangeThemeSizeSemanticTokensProvider``)
+    ///    - spaces: Provider for space semantic tokens. If nil, a default one will be used (``OrangeThemeSpaceSemanticTokensProvider``)
+    public init(sizes: AllSizeSemanticTokensProvider? = nil,
+                spaces: AllSpaceSemanticTokensProvider? = nil) {
         OL.debug("Init of OrangeThemeBreadcrumbComponentTokensProvider")
-        self.sizes = sizes
-        self.spaces = spaces
+        self.sizes = (sizes ?? OrangeThemeSizeSemanticTokensProvider())
+        self.spaces = (spaces ?? OrangeThemeSpaceSemanticTokensProvider())
     }
 
     deinit { }